+++ common/src/main/java/com/st0x0ef/stellaris/mixin/LivingEntityMixin.java
@@ -20,12 +20,12 @@ public abstract class LivingEntityMixin {
 
     @Inject(at = @At("HEAD"), method = "tick()V")
     private void tick(CallbackInfo ci){
-        //LivingEntity livingEntity = (LivingEntity) ((Object) this);
         ResourceKey<Level> dimension = stellaris$livingEntity.level().dimension();
+
         if (PlanetUtil.isPlanet(dimension)) {
-            stellaris$livingEntity.getAttribute(Attributes.GRAVITY).setBaseValue((double) Utils.MPS2ToMCG(PlanetUtil.getPlanet(dimension).gravity()));
+            stellaris$livingEntity.getAttribute(Attributes.GRAVITY).setBaseValue(Utils.MPS2ToMCG(PlanetUtil.getPlanet(dimension).gravity()));
         } else {
-            stellaris$livingEntity.getAttribute(Attributes.GRAVITY).setBaseValue((double) 0.08);
+            stellaris$livingEntity.getAttribute(Attributes.GRAVITY).setBaseValue(0.08);
         }
     }
 }
